src/server/requirements.txt
Flask
Flask-Script
Flask-Assets
Flask-pymongo
cssmin
gevent-socketio
pyjade
gevent_zeromq
itsdangerous
